Switch to composite projects
