+++ .gitignore
@@ -3,3 +3,5 @@ node_modules/
 .cache/
 *.log
 **/config.ts
+*.tsbuildinfo
+out/
