+++ .vscode/tasks.json
@@ -5,54 +5,25 @@
   "tasks": [
     {
       "type": "npm",
-      "script": "build-server",
-      "label": "Build server",
-      "problemMatcher": ["$tsc"],
-    },
-    {
-      "type": "typescript",
-      "label": "Watch server",
-      "tsconfig": "server/tsconfig.json",
-      "option": "watch",
-      "isBackground": true,
-      "problemMatcher": ["$tsc-watch"],
-      "presentation": {
-        "echo": true,
-        "reveal": "never",
-        "panel": "dedicated",
-        "showReuseMessage": false,
-        "clear": true,
-      },
-      "runOptions": {
-        "runOn": "folderOpen",
-      },
-    },
-    {
-      "type": "typescript",
-      "option": "watch",
-      "label": "Watch and typecheck extension",
-      "tsconfig": "extension/tsconfig.json",
-      "isBackground": true,
-      "problemMatcher": ["$tsc-watch"],
-      "presentation": {
-        "echo": true,
-        "reveal": "never",
-        "panel": "dedicated",
-        "showReuseMessage": false,
-        "clear": true,
-      },
-      "runOptions": {
-        "runOn": "folderOpen",
-      },
-    },
-    {
-      "type": "npm",
       "script": "serve-ext",
       "label": "Serve and watch extension",
       "problemMatcher": [],
       "presentation": {
         "reveal": "never",
         "panel": "dedicated",
+      },
+    },
+    {
+      "type": "npm",
+      "script": "watch-ts",
+      "problemMatcher": ["$tsc-watch"],
+      "group": "build",
+      "isBackground": true,
+      "presentation": {
+        "reveal": "never",
+      },
+      "runOptions": {
+        "runOn": "folderOpen",
       },
     },
   ],
