+++ Dockerfile
@@ -16,5 +16,5 @@ COPY ./ /srv
 
 WORKDIR /srv/server
 EXPOSE 80 443 8080
-CMD ["node", "--max_old_space_size=4096", "./dist/server.js"]
+CMD ["node", "--max_old_space_size=4096", "./out/server.js"]
 USER node
