+++ server/src/language-server.ts
@@ -13,10 +13,10 @@ import {
     MessageConnection,
 } from 'vscode-jsonrpc'
 import { LogMessageNotification } from 'vscode-languageserver-protocol'
-import { LangTypescriptConfiguration } from './config'
+import { LangTypescriptConfiguration } from '../../common/src/config'
+import { LOG_LEVEL_TO_LSP, Logger, LSP_TO_LOG_LEVEL, PrefixedLogger } from '../../common/src/logging'
 import { createDispatcher, Dispatcher } from './dispatcher'
 import { disposeAll, subscriptionToDisposable } from './disposable'
-import { LOG_LEVEL_TO_LSP, Logger, LSP_TO_LOG_LEVEL, PrefixedLogger } from './logging'
 
 const TYPESCRIPT_LANGSERVER_JS_BIN = path.resolve(
     __dirname,
