+++ server/src/tsconfig.ts
@@ -3,11 +3,11 @@ import { writeFile } from 'mz/fs'
 import { Span, Tracer } from 'opentracing'
 import { fileURLToPath, URL } from 'url'
 import { CancellationToken } from 'vscode-jsonrpc'
-import { throwIfCancelled } from './cancellation'
-import { flatMapConcurrent } from './ix'
-import { Logger } from './logging'
+import { throwIfCancelled } from '../../common/src/cancellation'
+import { flatMapConcurrent } from '../../common/src/ix'
+import { Logger } from '../../common/src/logging'
+import { logErrorEvent, tracePromise } from '../../common/src/tracing'
 import { ResourceRetrieverPicker } from './resources'
-import { logErrorEvent, tracePromise } from './tracing'
 
 export async function sanitizeTsConfigs({
     dir,
